pyproject.toml
[tool.poetry]
name = "test-sqlalchemy"
version = "0.1.0"
description = ""
authors = ["lavrik91 <lavrik2837@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
psycopg = "^3.1.13"
pydantic-settings = "^2.1.0"
psycopg-binary = "^3.1.13"
pytest = "^7.4.3"
pytest-dotenv = "^0.5.2"
psycopg2 = "^2.9.9"
pytest-mock = "^3.12.0"
mock-alchemy = "^0.2.6"
pika = "^1.3.2"
httpx = "^0.25.2"
pytest-asyncio = "^0.23.2"
gunicorn = "^21.2.0"
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
